pyproject.toml
[project]
name = "cricket-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb>=1.10.0",
    "sqlfluff>=3.5.0",
    "sqlfluff-templater-dbt>=3.5.0",
    "google-genai>=1.56.0",
    "duckdb>=1.1.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]
